more refactor - WIP
