+++ average_times_at_locations.csv
@@ -1,0 +1,14 @@
+,0
+Initial Triage/Water Pickup,4.2788265829653325
+Main Battle Dressing Station,0.0
+Holding Area,0.0
+Auxillary Treatment Area,0.0
+Other Location,0.0
+Waiting For Red Doctor,0.0
+Waiting For Yellow Doctor,0.0
+Waiting for Green Corpsman,0.0
+Waiting for Black Corpsman,0.0
+With Red Dedicated Doctor,19.427204639663472
+With Yellow Dedicated Doctor,18.9287131927502
+With Green Dedicated Corpsman,2.083416718992538
+With Black Dedicated Corpsman,0.0
