+++ individual_marines_priority_count.csv
@@ -1,5 +1,5 @@
-,0,1,2
-Priority 1: Red,6.0,,
-Priority 2: Yellow,4.0,5.0,
-Priority 3: Green,7.0,8.0,9.0
-Priority 4: Black,,,
+,0,1,2,3
+Priority 1: Red,7.0,,,
+Priority 2: Yellow,4.0,5.0,8.0,9.0
+Priority 3: Green,6.0,,,
+Priority 4: Black,,,,
