+++ individual_times_at_locations.csv
@@ -1,14 +1,14 @@
 ,0,1,2,3,4,5
-Initial Triage/Water Pickup,1.0848129112582647,14.085974867630362,1.897643363374371,5.633990879279352,4.97393222094999,3.5437757618651915
+Initial Triage/Water Pickup,2.1527079526314026,12.20641780476194,3.288513098341184,1.887498196690018,5.51787220609188,0.6199502392755676
 Main Battle Dressing Station,0.0,,,,,
-Holding Area,0.0,0.0,,,,
-Auxillary Treatment Area,0.0,0.0,0.0,,,
+Holding Area,0.0,0.0,0.0,0.0,,
+Auxillary Treatment Area,0.0,,,,,
 Other Location,,,,,,
 Waiting For Red Doctor,0.0,,,,,
-Waiting For Yellow Doctor,0.0,0.0,,,,
-Waiting for Green Corpsman,0.0,0.0,0.0,,,
+Waiting For Yellow Doctor,0.0,0.0,0.0,0.0,,
+Waiting for Green Corpsman,0.0,,,,,
 Waiting for Black Corpsman,,,,,,
-With Red Dedicated Doctor,34.23187730418556,,,,,
-With Yellow Dedicated Doctor,22.992916506939252,11.095969738454002,,,,
-With Green Dedicated Corpsman,18.685971527998134,5.511937613521411,32.481905994780846,,,
+With Red Dedicated Doctor,19.427204639663472,,,,,
+With Yellow Dedicated Doctor,26.172487142031173,13.352911652062886,14.564034053236526,21.625419923670222,,
+With Green Dedicated Corpsman,2.083416718992538,,,,,
 With Black Dedicated Corpsman,,,,,,
