+++ source/MassCasualtySystem/Mass_Casualty_System.py
@@ -1,192 +1,11 @@
-#IMPORTS
-import math
-import numpy
-from numpy import random
-from numpy import mean
-import scipy
-from scipy import stats
-from scipy.stats import lognorm
+
 import simpy
-import pandas
-import matplotlib
-from matplotlib import pyplot
+import numpy
 
-#Class representing...
-#VARIABLES AND PARAMETERS
-#Global variables and a parameters are tweaked here.
-class VariablesAndParameters:
-    #Sim Details
-    warm_up = 3
-    number_of_runs = 10
-    run_number = 0
-    duration_of_simulation_in_minutes = 500
-    simulation_time = (duration_of_simulation_in_minutes * 60)
-    #SIMULATION TIME IS IN SECONDS
-
-    #People Involved
-    #We are calculating how many Marines we can handle before resources are swamped or strained. Therefore, marine_counter starts and stays at 0.
-    marine_counter = 0
-    number_of_water_pickup_triage_personnel = 10
-    number_of_red_dedicated_doctors = 281
-    number_of_yellow_dedicated_doctors = 260
-    number_of_green_dedicated_corpsmen = 239
-    number_of_black_dedicated_corpsmen = 239
-
-    #Location Occupancy
-    main_bds_maximum_occupancy = 92
-    holding_area_maximum_occupancy = 280
-    auxillary_treatment_area_maximum_occupancy = 370
-    other_location_maximum_occupancy = 270
-
-    #How Long It Takes To Do Things
-    #Gonna want to mess around with a lognormal calculator and see what some acceptable samples could be.
-    interarrival_mean = 1
-    water_pickup_triage_mean_time = 1.2
-    water_pickup_triage_stdev_time = 1
-    red_mean_doctor_main_bds_consultation = 4
-    red_stdev_doctor_main_bds_consultation = 1
-    yellow_mean_doctor_holding_area_consultation = 3
-    yellow_stdev_doctor_holding_area_consultation = 1
-    green_mean_corpsman_auxillary_treatment_area_consultation = 2
-    green_stdev_corpsman_auxillary_treatment_area_consultation = 1
-    black_mean_corpsman_other_location_consultation = 2
-    black_stdev_corpsman_other_location_consultation = 1
-
-    #INSIGHTS
-    #How Long It Takes From Water to Color-Coded Area
-    wait_time_of_inital_triage_water_pickup = []
-    red_wait_time_between_triage_to_main_bds = []
-    yellow_wait_time_between_triage_to_holding_area = []
-    green_wait_time_between_triage_to_auxillary_treatment_area = []
-    black_wait_time_between_triage_to_other_location = []
-
-    #How Long It Takes Once In Color-Coded To Get A Provider
-    red_wait_for_doctor_main_bds = []
-    yellow_wait_for_doctor_holding_area = []
-    green_wait_for_corpsman_auxillary_treatment_area = []
-    black_wait_for_corpsman_other_location = []
-
-    #How Long It Takes In Color-Coded Area For Care
-    red_time_in_main_bds = []
-    yellow_time_in_holding_area = []
-    green_time_in_aux_treatment_area = []
-    black_time_in_other_location = []
-
-#Class representing...
-#A MARINE
-class InstanceOfMarine:
-    #ID and Time In System
-    def __init__(self):
-        self.time_in_system_start = 0
-        self.id = VariablesAndParameters.marine_counter
-        #DEBUGGING
-        print("ID: ", self.id)
-        print("Start Time In System: ", self.time_in_system_start)
-
-        #Give Them Priority Number and Color
-        self.set_priority()
-        self.set_triage_color()
-
-        VariablesAndParameters.marine_counter += 1
-
-        #DEBUGGING
-        print("Marine Counter: ", VariablesAndParameters.marine_counter)
-
-    #Priority Number To Correspond With Color
-    def set_priority(self):
-        self.set_priority_number = numpy.random.choice([1, 2, 3, 4])
-
-        #DEBUGGING
-        print("Priority Number: ", self.set_priority_number)
-
-    #Priority Color That Corresponds To Priority Number
-    def set_triage_color(self):
-        if self.set_priority_number == 1:
-            self.triage_color = "Red"
-            #Add To Data
-            if VariablesAndParameters.run_number > VariablesAndParameters.warm_up:
-                Track.individual_marines_priority_count_dictionary["Priority 1: Red"].append(self.id)
-            
-        if self.set_priority_number == 2:
-            self.triage_color = "Yellow"
-            #Add To Data
-            if VariablesAndParameters.run_number > VariablesAndParameters.warm_up:
-                Track.individual_marines_priority_count_dictionary["Priority 2: Yellow"].append(self.id)
-            
-        if self.set_priority_number == 3:
-            self.triage_color = "Green"
-            #Add To Data
-            if VariablesAndParameters.run_number > VariablesAndParameters.warm_up:
-                Track.individual_marines_priority_count_dictionary["Priority 3: Green"].append(self.id)
-            
-        if self.set_priority_number == 4:
-            self.triage_color = "Black"
-            #Add To Data
-            if VariablesAndParameters.run_number > VariablesAndParameters.warm_up:
-                Track.individual_marines_priority_count_dictionary["Priority 4: Black"].append(self.id)
-                
-        #DEBUGGING
-        print("Triage Color: ", self.triage_color)
-
-#Class representing...
-#DATA
-#Where we put data as the simulation runs.
-class Track:
-
-    #Store Individual IDs of Marines With Each Priority
-    #Used during execution for storage of priorities as they are assigned.
-    individual_marines_priority_count_dictionary = {
-        "Priority 1: Red": [],
-        "Priority 2: Yellow": [],
-        "Priority 3: Green": [],
-        "Priority 4: Black": []
-    }
-
-    #Store Number of Each Priority
-    #Used at the end to hold calculated values.
-    total_priority_count_dictionary = {
-        "Priority 1: Red": [],
-        "Priority 2: Yellow": [],
-        "Priority 3: Green": [],
-        "Priority 4: Black": []
-
-    }
-
-    #Store Individual Times at Location in a List
-    #Used during execution to store times of Marines at locations as they go through the system.
-    individual_times_at_locations_dictionary = {
-        "Initial Triage/Water Pickup": [],
-        "Main Battle Dressing Station": [],
-        "Holding Area": [],
-        "Auxillary Treatment Area": [],
-        "Other Location": [],
-        "Waiting For Red Doctor": [],
-        "Waiting For Yellow Doctor": [],
-        "Waiting for Green Corpsman": [],
-        "Waiting for Black Corpsman": [],
-        "With Red Dedicated Doctor": [],
-        "With Yellow Dedicated Doctor": [],
-        "With Green Dedicated Corpsman": [],
-        "With Black Dedicated Corpsman": []
-    }
-
-    #Store Average Time At Locations
-    #Used at the end for calculations.
-    average_times_at_locations_dictionary = {
-        "Initial Triage/Water Pickup": [],
-        "Main Battle Dressing Station": [],
-        "Holding Area": [],
-        "Auxillary Treatment Area": [],
-        "Other Location": [],
-        "Waiting For Red Doctor": [],
-        "Waiting For Yellow Doctor": [],
-        "Waiting for Green Corpsman": [],
-        "Waiting for Black Corpsman": [],
-        "With Red Dedicated Doctor": [],
-        "With Yellow Dedicated Doctor": [],
-        "With Green Dedicated Corpsman": [],
-        "With Black Dedicated Corpsman": []
-    }
+# user-defined classes
+from variables_and_parameters import VariablesAndParameters
+from marine import InstanceOfMarine
+from track import Track
 
 #Class representing...
 #MASS CASUALTY TRIAGE
@@ -585,153 +404,3 @@ class MassCasualtySystem:
                     Track.individual_times_at_locations_dictionary["With Black Dedicated Corpsman"].append(float(self.black_corpsman_care_elapsed_time))
 
                 VariablesAndParameters.run_number += 1
-
-#Class representing...
-#CALCULATIONS
-#The putting together the outputs of the gonkulator.
-class CalculationsAndConversions:
-    def total_priority_counts():
-        for key, values in Track.individual_marines_priority_count_dictionary.items():
-            Track.total_priority_count_dictionary[key] = len(values)
-        print("Total Priority Count Dictionary:")
-        print(Track.total_priority_count_dictionary.values())
-
-    def average_times_at_locations():
-        for key, values in Track.individual_times_at_locations_dictionary.items():
-            if len(values) > 0:
-                Track.average_times_at_locations_dictionary[key] = sum(values) / len(values)
-            else:
-                Track.average_times_at_locations_dictionary[key] = 0
-        print("Average Time At Location Dictionary:")
-        print(Track.average_times_at_locations_dictionary.values())
-
-    def convert_to_dataframe_individual_marines_priority_count():
-        """
-        Converting our Dictionaries to Pandas Dataframes
-        Here we are converting our dictionaries we used during the simulation to Pandas dataframes for easy output and plotting later.
-        Pandas dataframes don't just automatically fill in any empty cells if there isn't any values there like Excel does.
-        So, if there are 17 Red priority Marines, and only 4 Yellow Priority Marines, we need to fill in the remaining values
-        on the Yellow Priority Marines rows with designated empty values, else the dataframe will freak and not work.
-        """
-        #INDIVIDUAL MARINES PRIORITY COUNT DICTIONARY
-        #Step 1: Find the longest entry.
-        max_length_individual_marines_priority_count_dictionary = max(len(values) for values in Track.individual_marines_priority_count_dictionary.values())
-        """
-        for entry in Track.individual_marines_priority_count_dictionary:
-            max_length_individual_marines_priority_count_dictionary = max(len(entry))
-        """
-        #Step 2: Padd the list with any extra values to make all entries of equal length.
-        """
-        Here, we are making a new dictionary to include the padded (empty) cells.
-        Then for every key and value pair in our unpadded dictionary, we add an empty cell.
-        [] * (max_length... - len(values)) says add an empty cell ([]) for however many cells we have between the largest entry in the dictionary 
-        and how many we actually have.
-        Then we go ahead and add the values plus our new padding into the dictionary.
-        """
-        padded_individual_marines_priority_count_dictionary = {}
-        for key, values in Track.individual_marines_priority_count_dictionary.items():
-            padding = [] * (max_length_individual_marines_priority_count_dictionary - len(values))
-            padded_values = values + padding
-            padded_individual_marines_priority_count_dictionary[key] = padded_values
-        #Step 3: Convert to Pandas Dataframe
-        individual_marines_priority_count_dataframe = pandas.DataFrame.from_dict(padded_individual_marines_priority_count_dictionary, orient="index")
-        print(individual_marines_priority_count_dataframe)
-        #Step 4: Output Dataframe as a CSV for future use.
-        individual_marines_priority_count_dataframe.to_csv("individual_marines_priority_count.csv")
-        return individual_marines_priority_count_dataframe
-
-    def convert_to_dataframe_total_priority_count():
-        """
-        We don't have to do the padding here. Why? Because all values are the same for every row.
-        """
-        total_marines_priority_count_dataframe = pandas.DataFrame.from_dict(Track.total_priority_count_dictionary, orient="index")
-        print(total_marines_priority_count_dataframe)
-        total_marines_priority_count_dataframe.to_csv("total_marines_priority_count.csv")
-        return total_marines_priority_count_dataframe
-
-    def convert_to_dataframe_individual_times_at_locations():
-        max_length_individual_times_at_locations_dictionary = max(len(values) for values in Track.individual_times_at_locations_dictionary.values())
-        
-        padded_individual_times_at_locations_dictionary = {}
-        for key, values in Track.individual_times_at_locations_dictionary.items():
-            padding = [] * (max_length_individual_times_at_locations_dictionary - len(values))
-            padded_values = values + padding
-            padded_individual_times_at_locations_dictionary[key] = padded_values
-        
-        individual_times_at_locations_dataframe = pandas.DataFrame.from_dict(padded_individual_times_at_locations_dictionary, orient="index")
-        print(individual_times_at_locations_dataframe)
-        individual_times_at_locations_dataframe.to_csv("individual_times_at_locations.csv")
-        return individual_times_at_locations_dataframe
-
-    def convert_to_dataframe_average_times_at_locations():
-        """
-        This one is a little different because sometimes when you run the simulation, you could end up with a marine at
-        every location, in which case checking for max_length will error out because all values are the same, OR...
-        You could not end up with a Marine at every location, and you needed to padd the cell.
-        To check this, we first assume we don't need any padding, and calcualte the length of the first key's values to
-        serve as our first checking point. From there, if another entry has a different length, then we realize we need
-        to padd, and switch our is_padding_needed variable to Yes, or True. Then we break out of this loop to return to
-        the other stuff we needed to do, instead of wasting our time and checking all the other lengths, when we already
-        know we're gonna need to pad.
-        """
-        value_lengths = [1] * len(Track.average_times_at_locations_dictionary)
-        max_length = 1
-
-        if len(set(value_lengths)) == 1 and max_length == 1:
-            average_times_at_locations_dataframe = pandas.DataFrame.from_dict(Track.average_times_at_locations_dictionary, orient="index")
-            print(average_times_at_locations_dataframe)
-            average_times_at_locations_dataframe.to_csv("average_times_at_locations.csv")
-        else:
-            max_length_average_times_at_locations_dictionary = max_length
-    
-            padded_average_times_at_locations_dictionary = {}
-            for key, value in Track.average_times_at_locations_dictionary.items():
-                padded_average_times_at_locations_dictionary[key] = [value]
-
-            average_times_at_locations_dataframe = pandas.DataFrame.from_dict(padded_average_times_at_locations_dictionary, orient="index")
-            print(average_times_at_locations_dataframe)
-            average_times_at_locations_dataframe.to_csv("average_times_at_locations.csv")
-
-        return average_times_at_locations_dataframe
-
-    def plot_and_save_graph(dataframe, title, plot_type, image_filename):
-        if plot_type == "chart":
-            dataframe.plot()
-        elif plot_type == "histogram":
-            dataframe.hist()
-        elif plot_type == "pie":
-            dataframe.plot(kind="pie", subplots=True, legend=False)
-        elif plot_type == "scatter":
-            for idx, row in dataframe.iterrows():
-                pyplot.scatter([idx] * len(row), row)
-        else:
-            print("Invalid plot type!")
-        
-        pyplot.title(title)
-        pyplot.savefig(image_filename)
-     
-#RUNNING THE SYSTEM
-model = MassCasualtySystem()
-model.env.process(model.pickup_and_care_procedures())
-model.env.run(until=VariablesAndParameters.simulation_time)
-
-#APPLYING CALCULATIONS AND CONVERSIONS AND THEN GRAPHING EVERYTHING
-CalculationsAndConversions.total_priority_counts()
-CalculationsAndConversions.average_times_at_locations()
-"""
-Here we're going to create a dictionary with the names of each dataframe as well
-the function we wrote earlier to actually create the dataframe. This allows us to not
-only convert our tracking dictionaries to dataframes, but also creates an easy way to
-pass those dataframes to matplotlib for graphing.
-"""
-dataframes = {
-    "Individual Priority Counts By ID": (CalculationsAndConversions.convert_to_dataframe_individual_marines_priority_count, "chart", "individual_marines_priority_count.png"),
-    "Total Priority Counts": (CalculationsAndConversions.convert_to_dataframe_total_priority_count, "bar", "total_priority_count.png"),
-    "Individual Times at Locations by ID": (CalculationsAndConversions.convert_to_dataframe_individual_times_at_locations, "scatter", "individual_times_at_locations.png"),
-    "Average Times at Locations": (CalculationsAndConversions.convert_to_dataframe_average_times_at_locations, "bar","average_times_at_locations.png")
-}
-
-for title, (function, plot_type, image_filename) in dataframes.items():
-    print(title)
-    dataframe = function()
-    CalculationsAndConversions.plot_and_save_graph(dataframe, title, plot_type, image_filename)
