+++ total_marines_priority_count.csv
@@ -1,5 +1,5 @@
 ,0
 Priority 1: Red,1
-Priority 2: Yellow,2
-Priority 3: Green,3
+Priority 2: Yellow,4
+Priority 3: Green,1
 Priority 4: Black,0
